requirements.txt
# pip install -r requirements.txt

Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=3.4.2
Pillow
pytorch>=1.6.0
torchvision>=0.7.0
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
opencv_python

